1ª versión del tema de VARIABLES OK

+++ Variables/entrada y salida/VAR_entrada_02.py
@@ -1,5 +1,5 @@
 #BIEN --> el valor se multiplica como número, no como cadena.
-poder_de_goku = float(input('¿Cuál es el poder de Goku vs Nappa?: ' ))
+poder_de_goku = int(input('¿Cuál es el poder de Goku vs Nappa?: ' ))
 kaio_ken = 3
 
 print("El poder de Goku tras un (Kaio-Ken x" + str(kaio_ken) + ") es:")
